+++ lib/codeql.js
@@ -27,6 +27,7 @@ const fs_1 = __importDefault(require("fs"));
 const path_1 = __importDefault(require("path"));
 const exec_1 = require("@actions/exec");
 const yaml = __importStar(require("js-yaml"));
+const deserialize_1 = require("./deserialize");
 const download_1 = require("./download");
 const interpret_1 = require("./interpret");
 /**
@@ -124,18 +125,7 @@ async function getBqrsInfo(codeql, bqrs) {
     if (bqrsInfoOutput.exitCode !== 0) {
         throw new Error(`Unable to run codeql bqrs info. Exit code: ${bqrsInfoOutput.exitCode}`);
     }
-    return JSON.parse(bqrsInfoOutput.stdout, (_, value) => {
-        if (value && typeof value === "object") {
-            for (const k in value) {
-                if (k.match(/-./)) {
-                    const l = k.replace(/-./g, (x) => x[1].toUpperCase());
-                    value[l] = value[k];
-                    delete value[k];
-                }
-            }
-        }
-        return value;
-    });
+    return (0, deserialize_1.deserialize)(bqrsInfoOutput.stdout);
 }
 // Generates results.csv from the given bqrs file
 async function outputCsv(codeql, bqrs) {
